requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.5
django-redis==6.0.0
eventlet
et_xmlfile==2.0.0
Faker==37.5.3
gensim==4.3.3
idna==3.10
joblib==1.5.1
kombu==5.5.4
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
prompt_toolkit==3.0.51
pythainlp==5.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.4.0
requests==2.32.4
scikit-learn==1.7.1
scipy==1.13.1
six==1.17.0
smart_open==7.3.0.post1
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.3
mysqlclient==2.2.4
django-celery-beat==2.8.1
gunicorn==22.0.0
